cleared APIs by transferring low level methods into core module
